requirements.txt
hidapi==0.14.0.post4
vgamepad==0.1.0
PyQt5==5.15.10
